more UI testing, split tests into seperate files
